+++ src/lazyPages.js
@@ -33,3 +33,4 @@ export const VrijwilligersDashboard = lazy(() => import('@/pages/Vrijwilligers/V
 export const VrijwilligersExemptions = lazy(() => import('@/pages/Vrijwilligers/VrijwilligersExemptions'));
 export const VrijwilligersIva = lazy(() => import('@/pages/Vrijwilligers/VrijwilligersIva'));
 export const VrijwilligersDiensten = lazy(() => import('@/pages/Vrijwilligers/VrijwilligersDiensten'));
+export const Vrijwillig = lazy(() => import('@/pages/Vrijwillig/Vrijwillig'));